+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 1.0.0
 
+* Added a `dart-show-hover` command bound to `C-c ?` which displays information
+  about the Dart API under the cursor.
+
 * `dart-executable-path`'s default value is now set correctly even if the
   executable named `dart` on the user's path is a symlink or a wrapper script.